Update libcsv_parser.c
My general approach is to first check to see if the character is a multi-character character (so, escaped or a line ending).  If it is, then finish reading the character before continuing.

After that, check to see if we're ending a line (line ending), or a ending a cell (comma, or other delimiter).

When we see an un-escaped quote, we are not allowed to finish cells or lines until the closing quote has passed.