Update libcsv_parser.c
Gonna take a break for lunch.
I think this is covering all the basic cases.  Note things we COULD handle without difficulty but aren't because they're out of spec include:
Handling different separators
allowing CL or RF to end lines, not just CLRF

On a side note: Really?  Escaping characters with \ isn't part of the spec?  I think I want to leave that in... it's too useful.